requirements.txt
python-telegram-bot==13.4
bahire-hasab==0.2.4
pillow
datetime